refactor: :recycle: group label and group edit updates
