+++ locales/bn/filters.json
@@ -154,6 +154,7 @@
     "email": "ইমেইল",
     "group_role_title": "ব্যবহারকারীর গ্রুপের ভূমিকা",
     "institution_title": "ইনস্টিটিউট",
+    "name_of_user": "ব্যবহারকারীর নাম",
     "phone": "ফোন নম্বর",
     "profession_title": "পেশা",
     "role_title": "ভূমিকা",
