+++ locales/bn/group.json
@@ -9,8 +9,8 @@
     "founder": "প্রতিষ্ঠাতা",
     "moderator": "মডারেটর",
     "no_access": "কাস্টম ক্ষেত্র বা গ্রুপের নিয়ম আপডেট করতে দয়া করে প্রশাসক যোগাযোগ করুন",
-    "title": "গ্রুপ অ্যাডমিনিস্ট্রেটররা",
-    "updated": "গ্রুপ অ্যাডমিনিস্ট্রেটররা সফলভাবে আপডেট"
+    "title": "ব্যবহারকারীর ভূমিকা",
+    "updated": "গ্রুপ ব্যবহারকারীদের সফলভাবে আপডেট করা হয়েছে"
   },
   "confirm_closed_group": "এটি একটি বন্ধ গ্রুপ, সদস্যতা অনুরোধ একটি মডারেটর দ্বারা অনুমোদিত হতে হবে। যোগদান যোগ পাঠাতে ঠিক আছে টিপুন",
   "confirm_leave": "দল পরিত্যাগ করুন?",
