+++ locales/fr/filters.json
@@ -154,6 +154,7 @@
     "email": "e-mail",
     "group_role_title": "Rôles des groupes d'utilisateurs",
     "institution_title": "Institut",
+    "name_of_user": "Nom de l'utilisateur",
     "phone": "Numéro de téléphone",
     "profession_title": "Métier",
     "role_title": "Rôle",
