+++ locales/gu/filters.json
@@ -154,6 +154,7 @@
     "email": "ઇમેઇલ",
     "group_role_title": "વપરાશકર્તા જૂથ ભૂમિકાઓ",
     "institution_title": "સંસ્થા",
+    "name_of_user": "વપરાશકર્તાનું નામ",
     "phone": "ફોન નંબર",
     "profession_title": "વ્યવસાય",
     "role_title": "ભૂમિકા",
