+++ locales/gu/group.json
@@ -9,8 +9,8 @@
     "founder": "સ્થાપના કરનાર",
     "moderator": "મધ્યસ્થી",
     "no_access": "કસ્ટમ ક્ષેત્રો અથવા જૂથ નિયમોને અપડેટ કરવા માટે કૃપા કરીને એડમિનિસ્ટ્રેટરનો સંપર્ક કરો",
-    "title": "ગ્રુપ સંચાલકો",
-    "updated": "ગ્રુપ સંચાલકો ને સફળતાપૂર્વક અપડેટ કર્યા"
+    "title": "વપરાશકર્તા ભૂમિકાઓ",
+    "updated": "જૂથ વપરાશકર્તાઓ સફળતાપૂર્વક અપડેટ થયા"
   },
   "confirm_closed_group": "આ એક બંધ ગ્રુપ છે, સભ્યપદ વિનંતીને મધ્યસ્થી દ્વારા મંજૂર કરવાની જરૂર છે. જોડાવા માટે ઠીક દબાવો",
   "confirm_leave": "ગ્રુપ છોડી દો?",
