+++ locales/hi/filters.json
@@ -154,6 +154,7 @@
     "email": "ईमेल",
     "group_role_title": "उपयोगकर्ता समूह भूमिकाएँ",
     "institution_title": "संस्था",
+    "name_of_user": "उपयोगकर्ता का नाम",
     "phone": "फ़ोन नंबर",
     "profession_title": "पेशा",
     "role_title": "भूमिका",
