+++ locales/hi/group.json
@@ -9,8 +9,8 @@
     "founder": "संस्थापकों",
     "moderator": "मध्यस्थ",
     "no_access": "कस्टम फ़ील्ड या समूह नियमों को अपडेट करने के लिए कृपया व्यवस्थापक से संपर्क करें",
-    "title": "समूह प्रशासक",
-    "updated": "समूह प्रशासकों ने सफलतापूर्वक अद्यतन किया"
+    "title": "उपयोगकर्ता भूमिका",
+    "updated": "समूह उपयोगकर्ता सफलतापूर्वक अपडेट किए गए"
   },
   "confirm_closed_group": "यह एक बंद समूह है, सदस्यता अनुरोध को मॉडरेटर द्वारा अनुमोदित किया जाना है।अनुरोध में शामिल होने के लिए ओके दबाएं",
   "confirm_leave": "समूह छोड़ दें?",
