+++ locales/kn/filters.json
@@ -154,6 +154,7 @@
     "email": "ಇಮೇಲ್",
     "group_role_title": "ಬಳಕೆದಾರರ ಗುಂಪು ಪಾತ್ರಗಳು",
     "institution_title": "ಸಂಸ್ಥೆ",
+    "name_of_user": "ಬಳಕೆದಾರರ ಹೆಸರು",
     "phone": "ದೂರವಾಣಿ ಸಂಖ್ಯೆ",
     "profession_title": "ವೃತ್ತಿ",
     "role_title": "ಪಾತ್ರ",
