+++ locales/kn/group.json
@@ -9,8 +9,8 @@
     "founder": "ಸಂಸ್ಥಾಪಕರು",
     "moderator": "ಮಾಡರೇಟರ್ ಗಳು",
     "no_access": "ಕಸ್ಟಮ್ ಕ್ಷೇತ್ರಗಳು ಅಥವಾ ಗುಂಪು ನಿಯಮಗಳನ್ನು ನವೀಕರಿಸಲು ದಯವಿಟ್ಟು ನಿರ್ವಾಹಕರನ್ನು ಸಂಪರ್ಕಿಸಿ",
-    "title": "ಗುಂಪು ನಿರ್ವಾಹಕರು",
-    "updated": "ಗುಂಪು ನಿರ್ವಾಹಕರು ಯಶಸ್ವಿಯಾಗಿ ನವೀಕರಿಸಿದರು"
+    "title": "ಬಳಕೆದಾರರ ಪಾತ್ರಗಳು",
+    "updated": "ಗುಂಪು ಬಳಕೆದಾರರನ್ನು ಯಶಸ್ವಿಯಾಗಿ ನವೀಕರಿಸಲಾಗಿದೆ"
   },
   "confirm_closed_group": "ಇದು ಮುಚ್ಚಿದ ಗುಂಪು, ಸದಸ್ಯತ್ವದ ವಿನಂತಿಯನ್ನು ಮಾಡರೇಟರ್ ಅವರು ಅನುಮೋದಿಸಬೇಕಾಗಿದೆ. ಸೇರಬಯಸುವ ವಿನಂತಿಯನ್ನು ಕಳುಹಿಸಲು ಓಕೆ ಗುಂಡಿ ಒತ್ತಿರಿ",
   "confirm_leave": "ಗುಂಪನ್ನು ತೊರೆಯಬೇಕೆ?",
