+++ locales/mg/filters.json
@@ -154,6 +154,7 @@
     "email": "mailaka",
     "group_role_title": "Andraikitry ny Usergroup",
     "institution_title": "Institute",
+    "name_of_user": "Anaran'ny mpampiasa",
     "phone": "Nomeraon-telefaona",
     "profession_title": "asa",
     "role_title": "anjara asa",
