+++ locales/ml/filters.json
@@ -154,6 +154,7 @@
     "email": "ഇമെയിൽ",
     "group_role_title": "യൂസർഗ്രൂപ്പ് റോളുകൾ",
     "institution_title": "ഇൻസ്റ്റിറ്റ്യൂട്ട്",
+    "name_of_user": "ഉപയോക്താവിൻ്റെ പേര്",
     "phone": "ഫോൺ നമ്പർ",
     "profession_title": "തൊഴിൽ",
     "role_title": "പങ്ക്",
