+++ locales/ml/group.json
@@ -9,8 +9,8 @@
     "founder": "സ്ഥാപകർ",
     "moderator": "മോഡറേറ്റർമാർ",
     "no_access": "ഇഷ്ടാനുസൃത ഫീൽഡുകളോ ഗ്രൂപ്പ് നിയമങ്ങളോ അപ്ഡേറ്റ് ചെയ്യുന്നതിന് അഡ്മിനിസ്ട്രേറ്ററുമായി ബന്ധപ്പെടുക",
-    "title": "ഗ്രൂപ്പ് അഡ്മിനിസ്ട്രേറ്റർമാർ",
-    "updated": "ഗ്രൂപ്പ് അഡ്മിനിസ്ട്രേറ്റർമാർ വിജയകരമായി അപ്ഡേറ്റുചെയ്തു"
+    "title": "ഉപയോക്തൃ റോളുകൾ",
+    "updated": "ഗ്രൂപ്പ് ഉപയോക്താക്കൾ വിജയകരമായി അപ്ഡേറ്റ് ചെയ്തു"
   },
   "confirm_closed_group": "ഇതൊരു അടച്ച ഗ്രൂപ്പാണ്, അംഗത്വ അഭ്യർത്ഥന ഒരു മോഡറേറ്റർ അംഗീകരിക്കേണ്ടതുണ്ട്. ചേരാൻ അഭ്യർത്ഥന അയയ്ക്കാൻ ശരി അമർത്തുക",
   "confirm_leave": "സംഘം വിടുക?",
