+++ locales/mr/filters.json
@@ -154,6 +154,7 @@
     "email": "ईमेल",
     "group_role_title": "वापरकर्ता गट भूमिका",
     "institution_title": "संस्था",
+    "name_of_user": "वापरकर्त्याचे नाव",
     "phone": "फोन नंबर",
     "profession_title": "व्यवसाय",
     "role_title": "भूमिका",
