+++ locales/mr/group.json
@@ -9,8 +9,8 @@
     "founder": "संस्थापक",
     "moderator": "नियंत्रक",
     "no_access": "सानुकूल फील्ड किंवा ग्रुप नियम अद्यतनित करण्यासाठी प्रशासक संपर्क साधा",
-    "title": "गट प्रशासक",
-    "updated": "गट प्रशासकांनी यशस्वीरित्या अद्यतनित केले"
+    "title": "वापरकर्ता भूमिका",
+    "updated": "गट वापरकर्ते यशस्वीरित्या अद्यतनित केले"
   },
   "confirm_closed_group": "हे एक बंद गट आहे, सदस्यता विनंती मॉडरेटरद्वारे मंजूर करणे आवश्यक आहे. विनंती पाठविण्यासाठी ओके दाबा",
   "confirm_leave": "गट सोडा?",
