+++ locales/or/filters.json
@@ -154,6 +154,7 @@
     "email": "ଇମେଲ୍ |",
     "group_role_title": "ଉପଯୋଗକର୍ତ୍ତା ଗୋଷ୍ଠୀ ଭୂମିକା |",
     "institution_title": "ଅନୁଷ୍ଠାନ",
+    "name_of_user": "ଉପଯୋଗକର୍ତ୍ତାଙ୍କ ନାମ",
     "phone": "ଫୋନ୍ ନମ୍ବର",
     "profession_title": "ବୃତ୍ତି",
     "role_title": "ଭୂମିକା",
