+++ locales/or/group.json
@@ -9,8 +9,8 @@
     "founder": "ପ୍ରତିଷ୍ଠାତା |",
     "moderator": "ମୋଡ୍ ହେଡ୍ ରେର୍ସର",
     "no_access": "କଷ୍ଟମ୍ ଫିଲ୍ଡ କିମ୍ବା ଗୋଷ୍ଠୀ ନିୟମକୁ ଅପଡେଟ୍ କରିବାକୁ ଦୟାକରି ଆଡମିନିଷ୍ଟ୍ରେଟର୍ ସହିତ ଯୋଗାଯୋଗ କରନ୍ତୁ |",
-    "title": "ଗୋଷ୍ଠୀ ବ୍ୟବସ୍ଥାପକ",
-    "updated": "ଗୋଷ୍ଠୀ ଆଡମିନିଷ୍ଟ୍ରେଟରମାନେ ସଫଳତାର ସହିତ ଅପଡେଟ୍ କରନ୍ତି |"
+    "title": "ବ୍ୟବହାରକାରୀ ଭୂମିକା",
+    "updated": "ଗୋଷ୍ଠୀ ବ୍ୟବହାରକାରୀମାନେ ସଫଳତାର ସହିତ ଅଦ୍ୟତନ କଲେ |"
   },
   "confirm_closed_group": "ଏହା ଏକ ବନ୍ଦ ଗୋଷ୍ଠୀ, ସଦସ୍ୟତା ଅନୁରୋଧ ଏକ ମୋଡରେଟର୍ ଦ୍ୱାରା ଅନୁମୋଦନ କରିବାକୁ ପଡିବ | ମିଳିତ ଅନୁରୋଧ ପଠାଇବାକୁ ଓକେ ଦବାନ୍ତୁ |",
   "confirm_leave": "ଗୋଷ୍ଠୀ ଛାଡନ୍ତୁ?",
