+++ locales/pa/filters.json
@@ -154,6 +154,7 @@
     "email": "ਈ - ਮੇਲ",
     "group_role_title": "ਉਪਭੋਗਤਾ ਸਮੂਹ ਦੀਆਂ ਭੂਮਿਕਾਵਾਂ",
     "institution_title": "ਇੰਸਟੀਚਿਟ",
+    "name_of_user": "ਉਪਭੋਗਤਾ ਦਾ ਨਾਮ",
     "phone": "ਫੋਨ ਨੰਬਰ",
     "profession_title": "ਪੇਸ਼ਾ",
     "role_title": "ਭੂਮਿਕਾ",
