+++ locales/pa/group.json
@@ -9,8 +9,8 @@
     "founder": "ਸੰਸਥਾਪਕ",
     "moderator": "ਸੰਚਾਲਕ",
     "no_access": "ਕਸਟਮ ਫੀਲਡ ਜਾਂ ਸਮੂਹ ਦੇ ਨਿਯਮ ਨੂੰ ਅਪਡੇਟ ਕਰਨ ਲਈ ਕਿਰਪਾ ਕਰਕੇ ਪ੍ਰਬੰਧਕ ਨਾਲ ਸੰਪਰਕ ਕਰੋ",
-    "title": "ਸਮੂਹ ਪ੍ਰਬੰਧਕ",
-    "updated": "ਸਮੂਹ ਪ੍ਰਬੰਧਕ ਸਫਲਤਾਪੂਰਵਕ ਅਪਡੇਟ ਹੋਏ"
+    "title": "ਉਪਭੋਗਤਾ ਰੋਲ",
+    "updated": "ਸਮੂਹ ਉਪਭੋਗਤਾਵਾਂ ਨੂੰ ਸਫਲਤਾਪੂਰਵਕ ਅੱਪਡੇਟ ਕੀਤਾ ਗਿਆ"
   },
   "confirm_closed_group": "ਇਹ ਇੱਕ ਬੰਦ ਸਮੂਹ ਹੈ, ਮੈਂਬਰਸ਼ਿਪ ਬੇਨਤੀ ਨੂੰ ਇੱਕ ਸੰਚਾਲਕ ਦੁਆਰਾ ਮਨਜ਼ੂਰੀ ਦੇਣੀ ਪੈਂਦੀ ਹੈ. ਬੇਨਤੀ ਭੇਜਣ ਲਈ ਠੀਕ ਦਬਾਓ",
   "confirm_leave": "ਨੰਬਰ ਛੱਡੋ?",
