+++ locales/sw/filters.json
@@ -154,6 +154,7 @@
     "email": "barua pepe",
     "group_role_title": "Wajibu wa Kikundi cha Watumiaji",
     "institution_title": "Taasisi",
+    "name_of_user": "Jina la Mtumiaji",
     "phone": "Nambari ya simu",
     "profession_title": "Taaluma",
     "role_title": "Wajibu",
