+++ locales/sw/group.json
@@ -9,8 +9,8 @@
     "founder": "Waanzilishi",
     "moderator": "Wasimamizi",
     "no_access": "Kurekebisha mashamba ya desturi au sheria za kikundi tafadhali wasiliana na msimamizi",
-    "title": "Wakurugenzi wa kikundi",
-    "updated": "Watawala wa kikundi updated kwa mafanikio."
+    "title": "Majukumu ya Mtumiaji",
+    "updated": "Watumiaji wa Kikundi Wamesasishwa Kwa Mafanikio"
   },
   "confirm_closed_group": "Hii ni kikundi kilichofungwa, ombi la uanachama linapaswa kupitishwa na msimamizi. Bonyeza OK kutuma ombi la kujiunga.",
   "confirm_leave": "Acha Group?",
