+++ locales/ta/filters.json
@@ -154,6 +154,7 @@
     "email": "மின்னஞ்சல்",
     "group_role_title": "பயனர் குழு பாத்திரங்கள்",
     "institution_title": "நிறுவனம்",
+    "name_of_user": "பயனரின் பெயர்",
     "phone": "தொலைபேசி எண்",
     "profession_title": "தொழில்",
     "role_title": "பங்கு",
