+++ locales/ta/group.json
@@ -9,8 +9,8 @@
     "founder": "நிறுவனர்",
     "moderator": "இடையீட்டாளர்கள்",
     "no_access": "விருப்ப புலங்கள் சேர்க்க அல்லது குழு விதிகள் புதுப்பிக்க நிர்வாகியை தொடர்பு கொள்ளவும்",
-    "title": "குழு நிர்வாகிகள்",
-    "updated": "குழு நிர்வாகிகள் வெற்றிகரமாக மேம்படுத்தப்பட்டது"
+    "title": "பயனர் பாத்திரங்கள்",
+    "updated": "குழு பயனர்கள் வெற்றிகரமாக புதுப்பிக்கப்பட்டனர்"
   },
   "confirm_closed_group": "இது ஒரு வரையறுக்கப்பட்ட குழு, உறுப்பினராக இடையீட்டாளரால் அங்கீகரிக்கப்பட வேண்டும். கோரிக்கை அனுப்ப சரி அழுத்தவும்",
   "confirm_leave": "குழுவிலிருந்து விலகு?",
