+++ locales/te/filters.json
@@ -154,6 +154,7 @@
     "email": "ఇమెయిల్",
     "group_role_title": "వినియోగదారు సమూహ పాత్రలు",
     "institution_title": "ఇనిస్టిట్యూట్",
+    "name_of_user": "వినియోగదారు పేరు",
     "phone": "ఫోను నంబరు",
     "profession_title": "వృత్తి",
     "role_title": "పాత్ర",
