+++ locales/te/group.json
@@ -9,8 +9,8 @@
     "founder": "స్థాపకులు",
     "moderator": "మోడరేటర్లు",
     "no_access": "కస్టమ్ ఫీల్డ్స్ లేదా గుంపు నియమాలను నవీకరించడానికి దయచేసి నిర్వాహకుడిని సంప్రదించండి",
-    "title": "సమూహం నిర్వాహకులు",
-    "updated": "సమూహం నిర్వాహకులు విజయవంతంగా నవీకరించారు"
+    "title": "వినియోగదారు పాత్రలు",
+    "updated": "సమూహం వినియోగదారులు విజయవంతంగా నవీకరించబడ్డారు"
   },
   "confirm_closed_group": "ఇది ఒక క్లోజ్డ్ సమూహం, సభ్యత్వ అభ్యర్థన మోడరేటర్ ద్వారా ఆమోదించబడాలి. అభ్యర్థన చేరడానికి సరే నొక్కండి",
   "confirm_leave": "బృందాన్ని వదులు?",
