+++ src/components/pages/document/list/filters/name-of-user/index.tsx
@@ -1,5 +1,6 @@
 import { Spinner } from "@chakra-ui/react";
 import { selectStyles } from "@components/form/configs";
+import useDocumentFilter from "@components/pages/document/common/use-document-filter";
 import { axGetUsersByID, axUserFilterSearch } from "@services/user.service";
 import { MENU_PORTAL_TARGET } from "@static/constants";
 import debounce from "debounce-promise";
@@ -7,10 +8,8 @@ import useTranslation from "next-translate/useTranslation";
 import React, { useEffect, useState } from "react";
 import AsyncSelect from "react-select/async";
 
-import useSpeciesList from "../../use-species-list";
-
 export default function UserFilterInput({ filterKey }) {
-  const { filter, addFilter, removeFilter } = useSpeciesList();
+  const { filter, addFilter, removeFilter } = useDocumentFilter();
   const { t } = useTranslation();
   const [defaultValue, setDefaultValue] = useState<any[]>();
 
@@ -46,6 +45,7 @@ export default function UserFilterInput({ filterKey }) {
       onChange={handleOnChange}
       placeholder={t("filters:user.search")}
       styles={selectStyles}
+      formatOptionLabel={(option) => `${option.label} (${option.value})`}
     />
   ) : (
     <Spinner />
