+++ src/components/pages/observation/list/filters/name-of-user/index.tsx
@@ -45,6 +45,7 @@ export default function UserFilterInput({ filterKey }) {
       onChange={handleOnChange}
       placeholder={t("filters:user.search")}
       styles={selectStyles}
+      formatOptionLabel={(option) => `${option.label} (${option.value})`}
     />
   ) : (
     <Spinner />
