+++ src/components/pages/species/list/filter/name-of-user/index.tsx
@@ -1,6 +1,5 @@
 import { Spinner } from "@chakra-ui/react";
 import { selectStyles } from "@components/form/configs";
-import useUserList from "@components/pages/user/common/use-user-filter";
 import { axGetUsersByID, axUserFilterSearch } from "@services/user.service";
 import { MENU_PORTAL_TARGET } from "@static/constants";
 import debounce from "debounce-promise";
@@ -8,8 +7,10 @@ import useTranslation from "next-translate/useTranslation";
 import React, { useEffect, useState } from "react";
 import AsyncSelect from "react-select/async";
 
+import useSpeciesList from "../../use-species-list";
+
 export default function UserFilterInput({ filterKey }) {
-  const { filter, addFilter, removeFilter } = useUserList();
+  const { filter, addFilter, removeFilter } = useSpeciesList();
   const { t } = useTranslation();
   const [defaultValue, setDefaultValue] = useState<any[]>();
 
@@ -45,6 +46,7 @@ export default function UserFilterInput({ filterKey }) {
       onChange={handleOnChange}
       placeholder={t("filters:user.search")}
       styles={selectStyles}
+      formatOptionLabel={(option) => `${option.label} (${option.value})`}
     />
   ) : (
     <Spinner />
