+++ src/components/pages/user/list/user-list-data/filters/list.tsx
@@ -14,12 +14,12 @@ import React from "react";
 import EmailFilter from "./email";
 import InstituteFilter from "./institute";
 import Location from "./location/map-area";
+import UserFilter from "./name-of-user";
 import PhoneNumberFilter from "./phone";
 import ProfessionFilter from "./profession";
 import RolesFilter from "./role";
 import SexTypeFilter from "./sex-type";
 import TimeFilter from "./time";
-import UserFilter from "./user";
 import UserGroupFilter from "./user-group";
 import UserNameFilter from "./username";
 
@@ -72,7 +72,7 @@ export default function FiltersList() {
           <>
             <AccordionButton>
               <Box flex={1} textAlign="left">
-                {t("filters:user.title")}
+                {t("filters:user.name_of_user")}
               </Box>
               <AccordionIcon />
             </AccordionButton>
